[#1065] FRAGO-3 — 토글 주입 대상 확장과 딥링크 빌더 Domain 상향
- 부록 D 에 FRAGO-3 를 누적했다. FRAGO-2 가 대상을 둘로 적었으나 `TodayAndNextWidgetView` 도 같은 `TodoToggleButton` 임베드를 갖고 있어 셋째로 편입됐는데 문서에 안 남아 있었다
- 같은 단편명령으로 FRAGO-1 의 하향 자리를 개정했다. 빌더가 받는 입력이 전부 `String`·`EventTime` 이라 Domain 지식이고, 표시 모델을 switch 하는 어댑터만 `CalendarPresentation` 에 남는다
- Task 9 에 Composed 2종 스냅샷 케이스를 더했다. 기준 브랜치로는 이 케이스를 못 돌려 이번 이관 판정에는 못 쓰고 앞으로의 회귀 그물이다
