+++ Domain/Tests/Utils/AICommandEntryLinkTests.swift
@@ -1,6 +1,6 @@
 //
 //  AICommandEntryLinkTests.swift
-//  CalendarPresentationTests
+//  DomainTests
 //
 //  Created by sudo.park on 9/9/26.
 //  Copyright © 2026 com.sudo.park. All rights reserved.
@@ -9,7 +9,7 @@
 import Testing
 import Foundation
 
-@testable import CalendarPresentation
+@testable import Domain
 
 
 struct AICommandEntryLinkTests {
