+++ TodoCalendarApp/AppExtensions/Widget/Sources/Widgets/AICommandWidget/AICommandShortcutWidget.swift
@@ -10,7 +10,7 @@ import WidgetKit
 import SwiftUI
 import Extensions
 import CommonPresentation
-import CalendarPresentation
+import Domain
 import WidgetScenes
